+++ src/app/stores/index.ts
@@ -1,7 +1,5 @@
 // Store exports
 
-export * from './auth-store';
-export * from './config-store';
 export * from './edit-mode-store';
 export * from './error-store';
 export * from './home-assistant-store';